# lp-leilao/src/main/java/lp/leilao/entities/devices/Switch.java
package lp.leilao.entities.devices;
import io.micronaut.serde.annotation.Serdeable;
import jakarta.persistence.Column;
import jakarta.persistence.DiscriminatorValue;
import jakarta.persistence.Entity;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;

@EqualsAndHashCode(callSuper = true)
@Data
@Entity
@NoArgsConstructor
@AllArgsConstructor
@Serdeable
@DiscriminatorValue("SWITCH")
public class Switch extends ComputingDevice {

    @Column(name = "numberOfPorts")
    private Integer numberOfPorts;
    @Column(name = "firmwareVersion")
    private String firmwareVersion;
}