# lp-leilao/target/classes/META-INF/swagger/swagger.yml
openapi: 3.0.1
paths:
...
  /auctions/create:
    post:
      operationId: createAuction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auction'
        required: true
      responses:
        "201":
          description: createAuction 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auction'
...
  /auctions/{id}:
    get:
      operationId: getAuction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getAuction 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auction'
    delete:
      operationId: deleteAuction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteAuction 204 response
  /clients/create:
    post:
      operationId: createClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        required: true
      responses:
        "201":
          description: createClient 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
...
  /clients/{id}:
    get:
      operationId: getClient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getClient 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
    put:
      operationId: updateClient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        required: true
      responses:
        "200":
          description: updateClient 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
    delete:
      operationId: deletarClient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deletarClient 204 response
  /devices/create:
    post:
      operationId: createDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComputingDevice'
        required: true
      responses:
        "201":
          description: createDevice 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDevice'
  /devices/hubs/create:
    post:
      operationId: createHub
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hub'
        required: true
      responses:
        "201":
          description: createHub 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
...
  /devices/hubs/{id}:
    get:
      operationId: getHubs
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getHubs 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
    delete:
      operationId: deleteHub
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteHub 204 response
...
  /devices/monitors/create:
    post:
      operationId: createMonitor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Monitor'
        required: true
      responses:
        "201":
          description: createMonitor 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
...
  /devices/monitors/{id}:
    get:
      operationId: getMonitor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getMonitor 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
    delete:
      operationId: deleteMonitor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteMonitor 204 response
  /devices/notebooks/create:
    post:
      operationId: createNotebook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notebook'
        required: true
      responses:
        "201":
          description: createNotebook 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
...
  /devices/notebooks/{id}:
    get:
      operationId: getNotebook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getNotebook 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
    delete:
      operationId: deleteNotebook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteNotebook 204 response
  /devices/routers/create:
    post:
      operationId: createRouter
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Router'
        required: true
      responses:
        "201":
          description: createRouter 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
...
  /devices/routers/{id}:
    get:
      operationId: getRouter
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getRouter 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
    delete:
      operationId: deleteRouter
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteRouter 204 response
  /devices/switches/create:
    post:
      operationId: createSwitch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Switch'
        required: true
      responses:
        "201":
          description: createSwitch 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
...
  /devices/switches/{id}:
    get:
      operationId: getSwitches
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getSwitches 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDeviceDTO'
    delete:
      operationId: deleteSwitch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteSwitch 204 response
  /devices/{id}:
    get:
      operationId: getDevice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getDevice 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputingDevice'
    delete:
      operationId: deleteDevice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteDevice 204 response
  /financialInstitution/create:
    post:
      operationId: createFI
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinancialInstitution'
        required: true
      responses:
        "201":
          description: createFI 201 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialInstitution'
...
  /financialInstitution/{id}:
    get:
      operationId: getFI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getFI 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialInstitution'
    delete:
      operationId: deleteFI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: deleteFI 204 response
...
components:
  schemas:
    Auction:
      type: object
      properties:
        address:
          type: string
        finalDate:
          type: string
          format: date-time
        financialInstitution:
          type: string
        id:
          type: integer
          format: int64
        initialDate:
          type: string
          format: date-time
        numAuction:
          type: integer
          format: int32
...
    Client:
      type: object
      properties:
        cpf:
          type: string
        digitalCertificate:
          type: string
        email:
          type: string
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
    ComputingDevice:
      type: object
      properties:
        brand:
          type: string
        description:
          type: string
        deviceValue:
          type: number
          format: double
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        quantity:
          type: integer
          format: int32
    ComputingDeviceDTO:
      type: object
      properties:
        antenna:
          type: boolean
        brand:
          $ref: '#/components/schemas/ComputingDevice'
        description:
          $ref: '#/components/schemas/ComputingDevice'
        deviceValue:
          $ref: '#/components/schemas/ComputingDevice'
        firmwareVersion:
          type: string
        name:
          $ref: '#/components/schemas/ComputingDevice'
        numberOfPorts:
          type: integer
          format: int32
        ports:
          type: string
        quantity:
          $ref: '#/components/schemas/ComputingDevice'
        refreshRate:
          type: string
        screenSize:
          type: string
        specification:
          type: string
        volts:
          type: string
    FinancialInstitution:
      type: object
      properties:
        address:
          type: string
        cnpj:
          type: string
        id:
          type: integer
          format: int64
        name:
          type: string
    Hub:
      type: object
      allOf:
      - $ref: '#/components/schemas/ComputingDevice'
      - properties:
          ports:
            type: string
          volts:
            type: string
    Monitor:
      type: object
      allOf:
      - $ref: '#/components/schemas/ComputingDevice'
      - properties:
          refreshRate:
            type: string
          screenSize:
            type: string
...
    Notebook:
      type: object
      allOf:
      - $ref: '#/components/schemas/ComputingDevice'
      - properties:
          specification:
            type: string
...
    Router:
      type: object
      allOf:
      - $ref: '#/components/schemas/ComputingDevice'
      - properties:
          antenna:
            type: boolean
    Switch:
      type: object
      allOf:
      - $ref: '#/components/schemas/ComputingDevice'
      - properties:
          firmwareVersion:
            type: string
          numberOfPorts:
            type: integer
            format: int32
...